replaced all pygame files with pyqt5 files
