--- client/modules/README.md
@@ -1,5 +1,0 @@
-*Not my modules, these are some open source modules that i found on github.<br>
-<br>
-Credits:<br>
-&nbsp;&nbsp;&nbsp;&nbsp;
-pygame_textInput : [Naeero](https://github.com/Nearoo/pygame-text-input)<br>
